requirements.txt
annotated-types
anyio
beautifulsoup4
cachetools
certifi
charset-normalizer
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
lxml
ntscraper
oauthlib
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dotenv
python-telegram-bot
requests
requests-oauthlib
rsa
sniffio
socksio
soupsieve
tqdm
tweepy
typing-inspection
typing_extensions
uritemplate
urllib3
notion-client